Add floating point

+++ include/arch/i386/mm/pmm.h
@@ -7,7 +7,7 @@
 
 #define HIGHER_KERNEL_ADDRESS_LOAD 0x0
 
-extern uint32_t _mmngr_mem_size;
+extern double _mmngr_mem_size;
 
 typedef struct 
 {
